requirements.txt
active==0.0.5a0
amqp==5.0.6
asgiref==3.4.1
autopep8==1.7.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
catch==0.1.0
celery==5.2.1
celery-progress==0.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.1
cp==2020.12.3
crashtest==0.3.1
crispy-bootstrap5==0.7
cryptography==35.0.0
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
dj-rest-auth==2.2.4
dj-static==0.0.6
Django==3.2.8
django-allauth==0.54.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-datatables-view==1.19.1
django-easy-audit==1.3.3
django-environ==0.8.1
django-filter==21.1
django-glrm==1.1.3
django-graylog==0.7.0
django-guardian==2.4.0
django-localflavor==3.1
django-minio-backend==3.3.2
django-polymorphic==3.1.0
django-prettyjson==0.4.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.10
django-safedelete==1.2.0
django-stdimage==5.3.0
django-timezone-field==4.2.1
django-widget-tweaks==1.4.9
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-gis==0.18
djangorestframework-simplejwt==5.2.0
drf-extra-fields==3.4.1
drf-writable-nested==0.6.3
drf-yasg==1.20.0
ecdsa==0.17.0
filelock==3.3.1
fonttools==4.28.2
graypy==2.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflect==5.5.2
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
kiwisolver==1.3.2
kombu==5.2.2
MarkupSafe==2.1.1
matplotlib==3.5.0
minio==7.1.1
model-mommy==2.0.0
nome==0.1.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pastel==0.2.1
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.4.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycodestyle==2.9.1
pycosat==0.6.3
pycparser==2.20
PyJWT==2.4.0
pylev==1.4.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==0.26.1
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
response==0.5.0
rsa==4.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.29
sqlparse==0.4.2
standardjson==0.3.1
static3==0.7.0
toml==0.10.2
tomli==1.2.2
ua-parser==0.10.0
uri==2.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.9.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
wrapt==1.13.3
